+++ src/app/timesheet/timesheet/ag-grid-timesheet-value-editor/ag-grid-timesheet-value-editor.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { AgGridTimesheetValueEditorComponent } from './ag-grid-timesheet-value-editor.component';
+
+describe('AgGridTimesheetValueEditorComponent', () => {
+  let component: AgGridTimesheetValueEditorComponent;
+  let fixture: ComponentFixture<AgGridTimesheetValueEditorComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ AgGridTimesheetValueEditorComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AgGridTimesheetValueEditorComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});